requirements.txt
aiohttp==3.9.0b0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.11.17
chardet==3.0.4
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
discord-py-interactions==5.10.0
discord-typings==0.7.0
discord.py==2.3.2
distro==1.8.0
emoji==2.8.0
fonttools==4.45.0
frozenlist==1.4.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.1
matplotlib==3.8.2
multidict==6.0.4
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
rfc3986==1.5.0
scikit-learn==1.3.2
scipy==1.11.3
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
threadpoolctl==3.2.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
wheel==0.41.3
yarl==1.9.2
